+++ electronic_payments/electronic_payments/doctype/electronic_payment_settings/common.py
@@ -456,3 +456,65 @@ def get_party_details(doc):
 		return frappe._dict(
 			{"doctype": "Supplier", "name": doc.supplier, "description": doc.supplier_name}
 		)
+
+
+def state_label_lookup(state):
+	state_to_label = {
+		"AL": "Alabama (AL)",
+		"AK": "Alaska (AK)",
+		"AZ": "Arizona (AZ)",
+		"AR": "Arkansas (AR)",
+		"CA": "California (CA)",
+		"CO": "Colorado (CO)",
+		"CT": "Connecticut (CT)",
+		"DE": "Delaware (DE)",
+		"DC": "District of Columbia (DC)",
+		"FL": "Florida (FL)",
+		"GA": "Georgia (GA)",
+		"GU": "Guam (GU)",
+		"HI": "Hawaii (HI)",
+		"ID": "Idaho (ID)",
+		"IL": "Illinois (IL)",
+		"IN": "Indiana (IN)",
+		"IA": "Iowa (IA)",
+		"KS": "Kansas (KS)",
+		"KY": "Kentucky (KY)",
+		"LA": "Louisiana (LA)",
+		"ME": "Maine (ME)",
+		"MH": "Marshall Islands (MH)",
+		"MD": "Maryland (MD)",
+		"MA": "Massachusetts (MA)",
+		"MI": "Michigan (MI)",
+		"MN": "Minnesota (MN)",
+		"MS": "Mississippi (MS)",
+		"MO": "Missouri (MO)",
+		"MT": "Montana (MT)",
+		"NE": "Nebraska (NE)",
+		"NV": "Nevada (NV)",
+		"NH": "New Hampshire (NH)",
+		"NJ": "New Jersey (NJ)",
+		"NM": "New Mexico (NM)",
+		"NY": "New York (NY)",
+		"NC": "North Carolina (NC)",
+		"ND": "North Dakota (ND)",
+		"MP": "Northern Marianna Island (MP)",
+		"OH": "Ohio (OH)",
+		"OK": "Oklahoma (OK)",
+		"OR": "Oregon (OR)",
+		"PA": "Pennsylvania (PA)",
+		"PR": "Puerto Rico (PR)",
+		"RI": "Rhode Island (RI)",
+		"SC": "South Carolina (SC)",
+		"SD": "South Dakota (SD)",
+		"TN": "Tennessee (TN)",
+		"TX": "Texas (TX)",
+		"UT": "Utah (UT)",
+		"VT": "Vermont (VT)",
+		"VI": "Virgin Islands (VI)",
+		"VA": "Virginia (VA)",
+		"WA": "Washington (WA)",
+		"WV": "West Virginia (WV)",
+		"WI": "Wisconsin (WI)",
+		"WY": "Wyoming (WY)",
+	}
+	return state_to_label.get(state.upper(), "")
