+++ electronic_payments/electronic_payments/doctype/electronic_payment_settings/mercury.py
@@ -20,6 +20,7 @@ from electronic_payments.electronic_payments.doctype.electronic_payment_settings
 	get_payment_amount,
 	process_electronic_payment,
 	queue_method_as_admin,
+	state_label_lookup,
 )
 
 
@@ -603,65 +604,3 @@ def fetch_mercury_transactions(settings):
 	# TODO
 	settings = frappe._dict(json.loads(settings)) if isinstance(settings, str) else settings
 	return []
-
-
-def state_label_lookup(state):
-	state_to_label = {
-		"AL": "Alabama (AL)",
-		"AK": "Alaska (AK)",
-		"AZ": "Arizona (AZ)",
-		"AR": "Arkansas (AR)",
-		"CA": "California (CA)",
-		"CO": "Colorado (CO)",
-		"CT": "Connecticut (CT)",
-		"DE": "Delaware (DE)",
-		"DC": "District of Columbia (DC)",
-		"FL": "Florida (FL)",
-		"GA": "Georgia (GA)",
-		"GU": "Guam (GU)",
-		"HI": "Hawaii (HI)",
-		"ID": "Idaho (ID)",
-		"IL": "Illinois (IL)",
-		"IN": "Indiana (IN)",
-		"IA": "Iowa (IA)",
-		"KS": "Kansas (KS)",
-		"KY": "Kentucky (KY)",
-		"LA": "Louisiana (LA)",
-		"ME": "Maine (ME)",
-		"MH": "Marshall Islands (MH)",
-		"MD": "Maryland (MD)",
-		"MA": "Massachusetts (MA)",
-		"MI": "Michigan (MI)",
-		"MN": "Minnesota (MN)",
-		"MS": "Mississippi (MS)",
-		"MO": "Missouri (MO)",
-		"MT": "Montana (MT)",
-		"NE": "Nebraska (NE)",
-		"NV": "Nevada (NV)",
-		"NH": "New Hampshire (NH)",
-		"NJ": "New Jersey (NJ)",
-		"NM": "New Mexico (NM)",
-		"NY": "New York (NY)",
-		"NC": "North Carolina (NC)",
-		"ND": "North Dakota (ND)",
-		"MP": "Northern Marianna Island (MP)",
-		"OH": "Ohio (OH)",
-		"OK": "Oklahoma (OK)",
-		"OR": "Oregon (OR)",
-		"PA": "Pennsylvania (PA)",
-		"PR": "Puerto Rico (PR)",
-		"RI": "Rhode Island (RI)",
-		"SC": "South Carolina (SC)",
-		"SD": "South Dakota (SD)",
-		"TN": "Tennessee (TN)",
-		"TX": "Texas (TX)",
-		"UT": "Utah (UT)",
-		"VT": "Vermont (VT)",
-		"VI": "Virgin Islands (VI)",
-		"VA": "Virginia (VA)",
-		"WA": "Washington (WA)",
-		"WV": "West Virginia (WV)",
-		"WI": "Wisconsin (WI)",
-		"WY": "Wyoming (WY)",
-	}
-	return state_to_label.get(state.upper(), "")
